@@ -1,97 +1,48 @@
 {
-  "name": "app",
-  "version": "1.3.0",
+  "name": "app-update",
+  "version": "0.1.0",
   "private": true,
-  "scripts": {
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "test": "node scripts/test.js",
-    "lint": "npx eslint --cache . && npx prettier --check .",
-    "lint:fix": "npx eslint --fix . && npx prettier --write ."
-  },
-  "proxy": "http://uwpokerclub.herokuapp.com",
+  "proxy": "http://localhost:5000",
   "dependencies": {
-    "@babel/core": "7.17.9",
-    "@pmmmwh/react-refresh-webpack-plugin": "0.4.2",
-    "@svgr/webpack": "5.5.0",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "^26.6.0",
-    "babel-loader": "8.2.5",
-    "babel-plugin-named-asset-import": "^0.3.8",
-    "babel-preset-react-app": "^10.0.1",
-    "bfj": "^7.0.2",
-    "bootstrap": "^4.6.1",
-    "camelcase": "^6.3.0",
-    "case-sensitive-paths-webpack-plugin": "2.4.0",
-    "css-loader": "4.3.0",
-    "dotenv": "16.0.0",
-    "dotenv-expand": "8.0.3",
-    "file-loader": "6.2.0",
-    "fs-extra": "^10.1.0",
-    "html-webpack-plugin": "4.5.2",
-    "identity-obj-proxy": "3.0.0",
-    "jquery": "^3.6.0",
-    "mini-css-extract-plugin": "0.11.3",
-    "node-sass": "^4.14.1",
-    "optimize-css-assets-webpack-plugin": "5.0.8",
-    "pnp-webpack-plugin": "1.7.0",
-    "popper.js": "^1.16.1",
-    "postcss-flexbugs-fixes": "4.2.1",
-    "postcss-loader": "3.0.0",
-    "postcss-normalize": "8.0.1",
-    "postcss-preset-env": "6.7.1",
-    "postcss-safe-parser": "5.0.2",
-    "prompts": "2.4.2",
-    "react": "^17.0.2",
-    "react-app-polyfill": "^2.0.0",
-    "react-dev-utils": "^11.0.4",
-    "react-dom": "^17.0.2",
-    "react-refresh": "^0.8.3",
-    "react-select": "^5.3.1",
-    "react-router-dom": "^5.3.1",
-    "resolve": "1.22.0",
-    "resolve-url-loader": "^5.0.0",
-    "sass-loader": "8.0.2",
-    "semver": "7.3.7",
-    "style-loader": "1.3.0",
-    "terser-webpack-plugin": "4.2.3",
-    "ts-loader": "^8.4.0",
-    "ts-pnp": "1.2.0",
-    "url-loader": "4.1.1",
-    "web-vitals": "^0.2.4",
-    "webpack": "4.46.0",
-    "webpack-dev-server": "3.11.3",
-    "webpack-manifest-plugin": "2.2.0",
-    "workbox-webpack-plugin": "5.1.4"
-  },
-  "devDependencies": {
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^13.3.0",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^14.17.22",
-    "@types/react": "^17.0.29",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.3.1",
-    "@typescript-eslint/eslint-plugin": "^4.29.1",
-    "@typescript-eslint/parser": "^4.29.1",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-react-app": "^6.0.0",
-    "eslint-plugin-flowtype": "^5.10.0",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-jest": "^24.7.0",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-testing-library": "^4.12.4",
-    "eslint-webpack-plugin": "^2.5.4",
-    "jest": "^26.6.3",
-    "jest-circus": "^27.2.5",
-    "jest-resolve": "^27.2.5",
-    "jest-watch-typeahead": "^0.6.5",
-    "prettier": "2.2.1",
-    "typescript": "^4.4.4"
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/user-event": "^13.2.1",
+    "@types/jest": "^27.0.1",
+    "@types/node": "^16.7.13",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "bootstrap": "^5.1.3",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "react-select": "^5.3.1",
+    "sass": "^1.51.0",
+    "typescript": "^4.4.2",
+    "web-vitals": "^2.1.0"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
